Updated Files for fix to the exe file crashing
Fix for a crash that occurs if the File Name variable is left empty/blank when inputting user data. Added the check to the Folder Name variable as well since the program will fail to make the Folder if the variable holds no value; and while the Result return Err check will tell the user that, why do the extra work if it will fail in the end? So, there's an earlier error check for that. Lastly, added a Result return Err check for the file_write function as I realized I never set one up for that function.